+++ Runtime/Interface/ReelDialogueView.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Martian.Reel
+{
+    public abstract class ReelDialogueView : MonoBehaviour
+    {
+        public abstract void ViewUpdate(Dictionary<string, string> dialogueInfo);
+    }
+}