+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.1.0] - 2024-09-03
+### Extremely Basic but Working
+- Added Dialogue Nodes that wait for input before moving next.
+- Added Delay Nodes.
+- Added Reel Director, now you can move through the nodes one by one in a coroutine.
+- Added Reel Views, to display dialogue information. Any information can be added to dialogue information, as long as it fits in string form. In the example
+it can be shown to store things like the color variable of a Speakers name. This can be used for more than dialogue as well, for example portraits and their various emotions.
+- Various other smaller things, including a VERY simple example scene.
+
 ## [1.0.0] - 2024-09-02
 ### First Release
 - There is literally nothing in here yet.